Update Enemy Drop
CB - Enemy Drops now affect the UI when picked up

+++ Source/TowerDefense/Core_GameState.cpp
@@ -27,7 +27,7 @@ float ACore_GameState::GetPlayerCurrencyAmount()
 	return playerCurrencyAmount;
 }
 
-void ACore_GameState::UpdatePlayerCurrencyAmount(bool addCurrency, float amount)
+void ACore_GameState::UpdatePlayerCurrencyAmount(bool addCurrency, int amount)
 {
 	if (addCurrency)
 	{
